+++ src/pages/mypage/pages/sales-list/index.style.ts
