+++ src/queries/products/index.ts
@@ -1,9 +1,15 @@
-import { getProducts, getProductsCuration, getProductsDetail } from '@/api/products';
+import {
+  getProducts,
+  getProductsCuration,
+  getProductsDetail,
+  getProductsSalesList,
+} from '@/api/products';
 import {
   GetProductsParams,
   GetProductsDetailResponse,
   GetProductsResponse,
   GetProductsCurationResponse,
+  GetProductsSalesListResponse,
 } from '@/types/products';
 import { useQuery } from '@tanstack/react-query';
 import { QUERY_KEYS } from '@/queries/query-keys';
@@ -42,6 +48,19 @@ export const useGetProductsCurationQuery = () => {
         return await getProductsCuration();
       } catch {
         throw new Error('상품 목록을 가져오는 데 실패했습니다.');
+      }
+    },
+  });
+};
+
+export const useGetProductsSalesListQuery = () => {
+  return useQuery<GetProductsSalesListResponse, Error>({
+    queryKey: QUERY_KEYS.GET_SALES_LIST,
+    queryFn: async () => {
+      try {
+        return await getProductsSalesList();
+      } catch {
+        throw new Error('판매 내역 정보를 가져오는데 실패했습니다.');
       }
     },
   });
